Finally, vertex data is now stored in the mesh object
The entire project now uses tabs instead of 4 spaces.
Renderer now has a dedicated function for creating a VAO.
Wrapper function also exists to pass in vertex and index data as a vector (recommended way to do it as you need to supply array length).
It is 4 am and I am going to bed so renderer is quite messy but it functions so that's all the matters right now.