Model lights and better model render flow
The light cubes now use a model instead of a generated mesh cube. Models now cache the camera and their shader so rendering requires less pointers to be passed around. I found an issue with how textures are loaded causing only the first one to be used when using models.